Bug 1310263 - Merge ContentLocalizationObserver and ChromeLocalizationObserver into LocalizationObserver
