+++ docs/bindings.md
@@ -1,23 +1,5 @@
 <!-- Generated by documentation.js. Update this documentation by updating the source code. -->
 
-# ContentLocalizationObserver
-
-**Extends LocalizationObserver**
-
-The `ContentLocalizationObserver` is an extension of a `LocalizationObserver`
-class which is used for all Web content.
-
-This class is used for all HTML content translations.
-
-# ChromeLocalizationObserver
-
-**Extends LocalizationObserver**
-
-The `ChromeLocalizationObserver` is an extension of a `LocalizationObserver`
-class which additionally collects all XBL binding nodes for translation.
-
-This API is useful for chrome-privileged HTML and XUL in Gecko.
-
 # LocalizationObserver
 
 The `LocalizationObserver` class is responsible for localizing DOM trees.
