--- src/bindings/content.js
@@ -1,13 +1,0 @@
-import LocalizationObserver from './base';
-
-/**
- * The `ContentLocalizationObserver` is an extension of a `LocalizationObserver`
- * class which is used for all Web content.
- *
- * This class is used for all HTML content translations.
- */
-export default class ContentLocalizationObserver extends LocalizationObserver {
-  translateRootContent(root) {
-    return this.translateFragment(root);
-  }
-}
